% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DoubleGam.R
\name{DoubleGam}
\alias{DoubleGam}
\title{Double Generalized Additive Models}
\usage{
DoubleGam(formulaM, formulaG = NULL, family = "gaussian", data,
  startM = NULL, startG = NULL, method = "quasi", selection = "GCV",
  weights = NULL, control = DoubleGamControl(), trace = FALSE,
  scale = 1, conf.type = "bayesian", ...)
}
\arguments{
\item{formulaM}{the mean formula - typically of the type y ~ bsp(x1) + bsp(x2). Parametric models or mixtures of
parametric and non parametric bases can be specified}

\item{formulaG}{the dispersion formula - as  for formulaM, but no response variable to be specified.
\code{formulaG = ~ 1} corresponds to the estimation of the dispersion as a constant (i.e. standard GAM)}

\item{family}{as in glm, can be "gaussian", "poisson" or "Binomial"}

\item{data}{dataset where variables are stored (optional)}

\item{startM}{starting value for the coefficients needed to estimate the mean function}

\item{startG}{starting value for the coefficients needed to estimate the dispersion function}

\item{method}{can be "quasi" (the default) or "pseudo", according to whether deviance or pearson's residuals should be used
as response variables when fitting the dispersion function}

\item{selection}{the method used to select the smoothing parameter. can be "GCV" (the default), "AIC" or
"none", in which case no selection is done and a smoothing parameter value should be provided in the \code{bsp()} call.}

\item{weights}{as in glm -  to be used with care, since the data get always weigthed by the estimated variance function}

\item{control}{a list to control some behaviours of the estimation procedure, see {\link{DoubleGamControl}}}

\item{trace}{should a trace to follow the convergence be printed? Default is FALSE}

\item{scale}{similar to weigth, to be used if the dispersion function should be kept fixed}

\item{conf.type}{character describing how to compute confidence intervals for the smooth components; can "bayesian" (default) or "naive" as per Wood (2006)}
}
\value{
An object of the \code{\link{gamMD}} class.
If both the mean and the dispersion function are estimated the object will contain the following elements:
\code{converged}, \code{convVec}, \code{data}, \code{fitG}, \code{fitM}, \code{iter}, \code{relE}.
\code{fitM} contains information on the mean estimation procedure;
\code{fitG} contains information on the dispersion estimation procedure and has the same elements as \code{fitM}.

If only the mean function is estimated all the values of the \code{fitM} object are given in \code{gamMD} object.

The elements of \code{fitM} and \code{fitG} are:

\item{coefficients}{the estimated regression coefficients}
\item{fitted.values}{the fitted values of the model}
\item{desMat}{the full design matrix of the model}
\item{dims}{the dimension of the design matrix associated to each component}
\item{family}{family used for fitting the model - set to \code{Gamma} for \code{fitG}}
\item{linear.predictors}{estimated linear predictors for the model}
\item{deviance}{deviance residuals - in \code{fitM} these are be scaled by the estimated dispersion function, if present}
\item{residuals}{Pearson residuals}
\item{s.resid}{standardised Pearson residuals}
\item{y, yd}{response variable used in, respectively, the mean and diseprsion estimation procedure}
\item{converged}{logic indicator on whether the last inner iteration has converged}
\item{sm.p}{smoothing parameters used in the estimation procedure - either fixed or chosen}
\item{GCV}{Generalised Cross Validation value}
\item{AIC}{Akaike Information criterion value}
\item{yt}{the variable response variable transformed according to the link function and then centered}
\item{df}{overall equivalent sdegrees of freedom for the fit}
\item{vecdf}{a vector given the giving of freedom used by each covariate in the model}
\item{cov.coef}{covariance matrix of the coefficients}
\item{formula}{formula used}
\item{dimsP}{size of the parametric part of the model}
}
\description{
A function to estimate both the mean and the dispersion function using B-splines bases.
}
\author{
Ilaria Prosdocimi (ilapro@ceh.ac.uk)
}
\references{
Gijbels, Prosdocimi and Claeskens (2010), Nonparametric estimation of mean and dispersion functions in extended generalized linear models, Test, 19(3), 560-608, doi:10.1007/s11749-010-0187-1,

Gijbels and Prosdocimi (2011), Smooth estimation of mean and dispersion function in extended generalized additive models with application to Italian induced abortion data, Journal of Applied Statistics, 38(11), 2391-2411, doi:10.1080/02664763.2010.550039
}
\seealso{
\code{\link{DoubleGamControl}}, \code{\link{DoubleRobGam}}, \code{\link{gamMD}}
}

