% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DoubleGam.R
\name{DoubleGamControl}
\alias{DoubleGamControl}
\title{Setting the DoubleGam fitting defaults}
\usage{
DoubleGamControl(maxitM = 30, maxitG = 30, tol = 10^-5, acc = 5 * 10^-3,
  maxitOUT = 55, lambdaM = c(1e-04, 35500), lambdaG = c(1e-04, 13500))
}
\arguments{
\item{maxitM}{the maximum number of itearations allowed for each inner mean estimation procedures}

\item{maxitG}{the maximum number of itearations allowed for each inner dispersion estimation procedures}

\item{tol}{the level of accuarancy required for each inner mean and dispersion estimation procedure to converge}

\item{acc}{the level of accuracy required for the outer iteration to converge}

\item{maxitOUT}{the maximum number of outer iteration allowed in the Double estimation procedure}

\item{lambdaM}{range of acceptable smoothing parameter values for the mean estiamation procedure}

\item{lambdaG}{range of acceptable smoothing parameter values for the dispersion estiamation procedure}
}
\value{
a list of control parameters
}
\description{
This function is used internally in the DoubleGam fitting procedure.
The function's parameters control the numerical properties of the Double fitting procedure.
Care should be taken when changing the parameters default values, things can go wrong
(but sometimes fitting doesn't work well with the initial parameters, so you'll need to fiddle aroud with this)
}

