% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DoubleRobGam.R
\name{DoubleRobGam}
\alias{DoubleRobGam}
\title{#Double Robust Generalized Additive Models
#
#A function to robustly estimate both the mean and the dispersion function using B-splines bases.
#The robustification is done in a similar fashion to the robustbase::glmrob function.
#
#}
\usage{
DoubleRobGam(formulaM, formulaG = NULL, family = "gaussian", data,
  startM = NULL, startG = NULL, selection = "RGCV", weights = NULL,
  control = DoubleRobGamControl(), trace = FALSE, scale = 1,
  weights.on.x = "none", method = "quasi", ...)
}
\arguments{
\item{formulaM}{the mean formula - typically of the type y ~ bs(x1) + bs(x2). Parametric models or mixtures of
#parametric and non parametric bases can be specified
#}

\item{formulaG}{the dispersion formula - as  for formulaM, but no response variable to be specified.
#\code{formulaG = ~ 1} corresponds to the estimation of the dispersion as a constant (i.e. standard GAM)
#}

\item{family}{as in glm, can be "gaussian", "poisson" or "Binomial"
#}

\item{data}{dataset where variables are stored (optional)
#}

\item{startM}{starting value for the coefficients needed to estimate the mean function
#}

\item{startG}{starting value for the coefficients needed to estimate the dispersion function
#}

\item{selection}{the method used to select the smoothing parameter.
#can be "RGCV" (the default), "RAIC", "GCV", "AIC" or "none", in which case no selection is done and a smoothing parameter value should be provided in the \code{bs()} call.
#The RGCV and RAIC are robustified version of GCV and AIC (see Croux et al. (2012))
#}

\item{weights}{as in glm - to be used with care, since the data get always weigthed by the estimated variance function
#}

\item{control}{a list to control some behaviours of the estimation procedure, see {\link{DoubleRobGamControl}}
#}

\item{trace}{should a trace to follow the convergence be printed? Default is FALSE
#}

\item{scale}{similar to weigth, to be used if the dispersion function should be kept fixed
#}

\item{weights.on.x}{mutated from robustbase::glmrob, a vector of weights for the x-variables. Used to accomodated for leverage points.
#
#
#}

\item{method}{can be "quasi" (the default) or "pseudo", according to whether deviance or pearson's residuals should be used
#as response variables when fitting the dispersion function
#}
}
\value{
An object of the \code{\link{gamMD}} class.
#If both the mean and the dispersion function are estimated the object will contain the following elements:
#\code{converged}, \code{convVec}, \code{data}, \code{fitG}, \code{fitM}, \code{iter}, \code{relE}.
#\code{fitM} contains information on the mean estimation procedure;
#\code{fitG} contains information on the dispersion estimation procedure and has the same elements as \code{fitM}.
#
#If only the mean function is estimated all the values of the \code{fitM} object are given in \code{gamMD} object.
#
#The elements of \code{fitM} and \code{fitG} are:
#
#\item{coefficients}{the estimated regression coefficients}
#\item{fitted.values}{the fitted values of the model}
#\item{desMat}{the full design matrix of the model}
#\item{dims}{the dimension of the design matrix associated to each component}
#\item{family}{family used for fitting the model - set to \code{Gamma} for \code{fitG}}
#\item{linear.predictors}{estimated linear predictors for the model}
#\item{deviance}{deviance residuals - in \code{fitM} these are be scaled by the estimated dispersion function, if present}
#\item{residuals}{Pearson residuals}
#\item{s.resid}{standardised Pearson residuals}
#\item{y, yd}{response variable used in, respectively, the mean and diseprsion estimation procedure}
#\item{converged}{logic indicator on whether the last inner iteration has converged}
#\item{sm.p}{smoothing parameters used in the estimation procedure - either fixed or chosen}
#\item{GCV}{Generalised Cross Validation value}
#\item{AIC}{Akaike Information criterion value}
#\item{RGCV}{Roubust Generalised Cross Validation value}
#\item{RAIC}{Roubust Akaike Information criterion value}
#\item{w.x}{the weights on the covariates used to correct for the effect of leverage points}
#\item{estRphi}{roubustly estimated dispersion parameter}
#\item{tcc}{tuning constant c used in the robust estimation}
#\item{yt}{the variable response variable transformed according to the link function and then centered}
#\item{df}{overall equivalent degrees of freedom for the fit}
#\item{vecdf}{a vector giving the degrees of freedom used by each covariate in the model}
#\item{cov.coef}{covariance matrix of the coefficients}
#\item{formula}{formula used}
#\item{dimsP}{size of the parametric part of the model}
#
#
#
}
\description{
#Double Robust Generalized Additive Models
#
#A function to robustly estimate both the mean and the dispersion function using B-splines bases.
#The robustification is done in a similar fashion to the robustbase::glmrob function.
#
#
}
\author{
Ilaria Prosdocimi (ilapro@ceh.ac.uk)
#
}
\references{
Croux, C., Gijbels, I. and Prosdocimi, I. (2012), Robust Estimation of Mean and Dispersion Functions in Extended Generalized Additive Models. Biometrics, 68: 31-44. doi: 10.1111/j.1541-0420.2011.01630.x
}
\seealso{
\code{\link{DoubleRobGamControl}}, \code{\link{DoubleGam}}, \code{\link{gamMD}}
#
}

