% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DoubleRobGam.R
\name{DoubleRobGamControl}
\alias{DoubleRobGamControl}
\title{Setting the DoubleRobGam fitting defaults}
\usage{
DoubleRobGamControl(maxitM = 35, maxitG = 35, tol = 10^-7, acc = 5 *
  10^-3, maxitOUT = 60, tccM = 1.345, tccG = 1.345, lambdaM = c(1e-04,
  3500), lambdaG = c(1e-04, 3500))
}
\arguments{
\item{maxitM}{the maximum number of iterations allowed for each inner mean estimation procedures}

\item{maxitG}{the maximum number of iterations allowed for each inner dispersion estimation procedures}

\item{tol}{the level of accuracy required for each inner mean and dispersion estimation procedure to converge}

\item{acc}{the level of accuracy required for the outer iteration to converge}

\item{maxitOUT}{the maximum number of outer iteration allowed in the Double estimation procedure}

\item{tccM}{the tuning constant c used in the robust estimation of the mean function}

\item{tccG}{the tuning constant c used in the robust estimation of the dispersion function}

\item{lambdaM}{range of acceptable smoothing parameter values for the mean estimation procedure}

\item{lambdaG}{range of acceptable smoothing parameter values for the dispersion estimation procedure}
}
\value{
a list of control parameters
}
\description{
This function is used internally in the DoubleRobGam fitting procedure.
The function's parameters control the numerical properties of the Double Robust fitting procedure.
Care should be taken when changing the parameters default values, things can go wrong
(but sometimes fitting doesn't work well with the initial parameters, so you'll need to fiddle around with this)
}

