% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gamMD.R
\name{gamMD}
\alias{gamMD}
\alias{plot.gamMD}
\alias{print.gamMD}
\alias{summary.gamMD}
\title{Double Generalised Additive Models class}
\usage{
\method{print}{gamMD}(object)

\method{summary}{gamMD}(object, ...)

\method{plot}{gamMD}(object, col = 1, xlab = NULL, ylab = "yt",
  one = TRUE, conf.level = 0.05, ci.plot = TRUE, ...)
}
\arguments{
\item{object}{an object of the gamMD class}

\item{col}{colour in which the lines in the plot should be drawn}

\item{xlab,ylab}{optional axis labs}

\item{one}{logic value to indicate whether mean and disperion figures should be included togther in one plot}

\item{conf.level}{confidence level at which confidence bands should be drawn (default is 0.05)}

\item{ci.plot}{logic value to indicate whether confidence bands should be drawn}
}
\description{
Typically the result of a DoubleGam or DoubleRobGam fit. Method functions as \code{summary}, \code{print} and \code{plot}.
Depending on whether the mean only or both the mean and the dispersion function are estimated the results will be different.
In both cases the information on the convergence status og the overall estimation is given by \code{converged}.
\code{iter} gives the total number of (outer) iterations;
\code{data} contains the dataset to which the model was fitted, if given.

If both the mean and the dispersion function are estimated the object will contain the following elements:
\code{converged}, \code{convVec}, \code{data}, \code{fitG}, \code{fitM}, \code{iter}, \code{relE}.
\code{fitM} contains information on the mean estimation procedure;
\code{fitG} contains information on the dispersion estimation procedure and has the same elements as \code{fitM}.

If only the mean function is estimated all the values of the \code{fitM} object are given in \code{gamMD} object.
}
\seealso{
\code{\link{DoubleRobGam}}, \code{\link{DoubleGam}}
}

