% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{rdoublebinom}
\alias{rdoublebinom}
\title{Random generator of Double Exponantial Binomial-type data (counts)}
\usage{
rdoublebinom(size, n, prob, theta)
}
\arguments{
\item{size}{sample size needed}

\item{n}{number of trials (called \code{size} in \code{rbinom})}

\item{prob}{vector of probabilities}

\item{theta}{dispersion parameter/function. \code{theta = 1} (the default) corresponds to the Binomial distribution}
}
\description{
Random generator of Double Exponantial Binomial-type data (counts)
}

