% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{rdoublepois}
\alias{rdoublepois}
\title{Random generator of Double Exponantial Poisson-type data (counts)}
\usage{
rdoublepois(size = 1, lambda, gamma = 1)
}
\arguments{
\item{size}{sample size needed}

\item{lambda}{mean parameter/function}

\item{gamma}{dispersion parameter/function. \code{gamma = 1} (the default) corresponds to the Poisson distribution}
}
\description{
Random generator of Double Exponantial Poisson-type data (counts)
}

